+++ changelog.md
@@ -3,6 +3,7 @@
     * Fixed issue where some packages were loading examples when examples were excluded.
     * Work on HTI-2 proposed US Core Subscription testing.
     * Added support for `Resource.meta.tag` in `ParsedSubscription` objects.
+    * Fixed issue where subscription status changes were not being reflected in the resource.
 
 * v2025.210
     * Updated NuGet dependencies: